fix!: Additional statuses were added, and the project structure was refactored
