+++ api/src/main/java/com/vitrum/api/credentials/authentication/AuthService.java
@@ -4,9 +4,9 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.vitrum.api.dto.Request.AuthenticationRequest;
 import com.vitrum.api.dto.Response.AuthenticationResponse;
 import com.vitrum.api.dto.Request.RegisterRequest;
-import com.vitrum.api.token.Token;
-import com.vitrum.api.token.TokenRepository;
-import com.vitrum.api.token.TokenType;
+import com.vitrum.api.credentials.token.Token;
+import com.vitrum.api.credentials.token.TokenRepository;
+import com.vitrum.api.credentials.token.TokenType;
 import com.vitrum.api.credentials.user.Role;
 import com.vitrum.api.credentials.user.User;
 import com.vitrum.api.credentials.user.UserRepository;
