+++ api/src/main/java/com/vitrum/api/credentials/token/TokenRepository.java
@@ -1,4 +1,4 @@
-package com.vitrum.api.token;
+package com.vitrum.api.credentials.token;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
