+++ api/src/main/java/com/vitrum/api/credentials/token/TokenType.java
@@ -1,0 +1,5 @@
+package com.vitrum.api.credentials.token;
+
+public enum TokenType {
+    BEARER
+}
