+++ api/src/main/java/com/vitrum/api/manager/team/TeamController.java
@@ -1,4 +1,4 @@
-package com.vitrum.api.team;
+package com.vitrum.api.manager.team;
 
 import com.vitrum.api.dto.Request.TeamCreationRequest;
 import lombok.RequiredArgsConstructor;
