+++ api/src/main/java/com/vitrum/api/manager/team/TeamRepository.java
@@ -1,4 +1,4 @@
-package com.vitrum.api.team;
+package com.vitrum.api.manager.team;
 
 import org.springframework.data.jpa.repository.JpaRepository;
