+++ api/src/main/java/com/vitrum/api/manager/team/TeamService.java
@@ -1,10 +1,10 @@
-package com.vitrum.api.team;
+package com.vitrum.api.manager.team;
 
 import com.vitrum.api.dto.Response.*;
-import com.vitrum.api.member.MemberRepository;
+import com.vitrum.api.manager.member.MemberRepository;
 import com.vitrum.api.dto.Request.TeamCreationRequest;
 import com.vitrum.api.credentials.user.User;
-import com.vitrum.api.member.RoleInTeam;
+import com.vitrum.api.manager.member.RoleInTeam;
 import com.vitrum.api.util.Converter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.dao.DataIntegrityViolationException;
