--- api/src/main/java/com/vitrum/api/token/TokenType.java
@@ -1,5 +1,0 @@
-package com.vitrum.api.token;
-
-public enum TokenType {
-    BEARER
-}
