+++ api/src/main/java/com/vitrum/api/util/Converter.java
@@ -2,8 +2,8 @@ package com.vitrum.api.util;
 
 import com.vitrum.api.credentials.user.User;
 import com.vitrum.api.dto.Response.*;
-import com.vitrum.api.member.Member;
-import com.vitrum.api.team.Team;
+import com.vitrum.api.manager.member.Member;
+import com.vitrum.api.manager.team.Team;
 import org.springframework.stereotype.Service;
 import java.util.List;
 import java.util.stream.Collectors;
